Move shared logic into a separate file to keep Extension API code out of the injected script
