+++ src/utils/storage.ts
@@ -1,17 +1,5 @@
-import * as v from "valibot";
 import { storage } from "#imports";
-
-const liveChatBehaviorSchema = v.picklist([
-  "followDefault",
-  "forceCollapsed",
-  "forceExpanded",
-]);
-
-export type LiveChatBehavior = v.InferOutput<typeof liveChatBehaviorSchema>;
-
-export function isLiveChatBehavior(value: unknown): value is LiveChatBehavior {
-  return v.safeParse(liveChatBehaviorSchema, value).success;
-}
+import type { LiveChatBehavior } from "./types";
 
 export const liveChatBehaviorStorage = storage.defineItem<LiveChatBehavior>(
   "local:liveChatBehavior",
